requirements.txt
flask
pytesseract
pillow
googletrans==4.0.0-rc1
